add db init script and app entry
