+++ backend/src/config/sentry.ts
@@ -3,14 +3,15 @@ import { logger } from './logger';
 
 let Sentry: any = null;
 if (env.sentryDsn) {
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    Sentry = require('@sentry/node');
-    Sentry.init({ dsn: env.sentryDsn, tracesSampleRate: 0.1 });
-  } catch (e) {
-    logger.warn('SENTRY_DISABLED');
-    Sentry = null;
-  }
+  import('@sentry/node')
+    .then((m) => {
+      Sentry = (m as any).default || m;
+      Sentry.init({ dsn: env.sentryDsn, tracesSampleRate: 0.1 });
+    })
+    .catch(() => {
+      logger.warn('SENTRY_DISABLED');
+      Sentry = null;
+    });
 }
 
 export { Sentry };
