+++ backend/src/lib/prisma.ts
@@ -1,9 +1,10 @@
+import { PrismaClient, Prisma as PrismaNS } from '@prisma/client';
 import { logger } from '../config/logger';
 import { env } from '../config/env';
 
 let prisma: any = null;
 let dbAvailable = false;
-let Prisma: any = { PrismaClientKnownRequestError: class extends Error { code = ''; } };
+let Prisma: any = PrismaNS;
 
 if (process.env.NODE_ENV === 'production' && env.offlineMode && !env.productionOfflineAllowed) {
   logger.error('Offline mode is not allowed in production');
@@ -18,10 +19,7 @@ if (env.prismaEnginesChecksumIgnore) {
 }
 
 try {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const { PrismaClient, Prisma: PrismaNS } = require('@prisma/client');
   prisma = new PrismaClient();
-  Prisma = PrismaNS;
   dbAvailable = true;
 } catch (e) {
   const msg = 'Prisma indisponible: ' + (e as Error).message;
